Backgroun Jobs instruction fix
su - openproject -c "bash -l" resulted on error "bash: cannot set terminal process group (-1): Inappropriate ioctl for device".
(more about that here http://serverfault.com/questions/411307/cannot-set-terminal-process-group-during-su-to-another-user-as-login-shell    )
su openproject --login is used generally on this instruction so harmonizing to that way of su usage

RVM is not used on on this guide so rake is not found on that version specific path. Changing to bundle which is widely used here elsewhere.